little error

+++ Code/P-O-CW/src/main/java/be/kuleuven/cs/robijn/autopilot/image/Image.java
@@ -345,7 +345,7 @@ public class Image {
 				float angleCos = (float) Math.sqrt(1 / (1 + Math.pow(ratio, 2)));
 				float planeAngle = (float) Math.acos(angleCos);
 				result = 1.32279f * (float) (0.5/Math.tan(degreesToRadians(angle)) + 0.5 / Math.sin(planeAngle));
-				if (!Double.isFinite(result))
+				if (!Float.isFinite(result))
 					return 100;
 				else
 					return result;
@@ -354,7 +354,7 @@ public class Image {
 				float pixels = getMaximumDistanceSpherePixels(hue, sat);
 				float angle = (pixels * getHorizontalAngle()) / getnbColumns();
 				result = 1.32279f * (float) (Math.sqrt(0.75) / Math.tan((degreesToRadians(angle))));
-				if (!Double.isFinite(result))
+				if (!Float.isFinite(result))
 					return 100;
 				else
 					return result;
